# Conflicts:
#	src/main/java/now/planit/Controllers/SubtaskController.java
#	src/main/java/now/planit/Controllers/TaskController.java
#	src/main/java/now/planit/Data/Repo/DBFacade.java
#	src/main/java/now/planit/Data/Repo/RepoSubtask.java
#	src/main/resources/templates/project/createSubtask.html
#	src/main/resources/templates/project/createTask.html
#	target/classes/templates/project/createSubtask.html
#	target/classes/templates/project/createTask.html

+++ src/main/java/now/planit/Controllers/ProjectController.java
@@ -1,10 +1,8 @@
 package now.planit.Controllers;
 
 import now.planit.Domain.Models.Project;
-import now.planit.Domain.Models.Task;
 import now.planit.Domain.Models.User;
 import now.planit.Domain.Services.ProjectService;
-import now.planit.Domain.Services.TaskService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,19 +17,14 @@ public class ProjectController {
     User user;
     ProjectService projectService = new ProjectService();
     ArrayList<Project> projects = new ArrayList<>();
-    TaskService taskService = new TaskService();
-    ArrayList<Task> tasks = new ArrayList();
-    String projectName;
 
 
     @GetMapping("/myProjects")
     public String myProjects(Model model, WebRequest request) {
-        user = (User) request.getAttribute("user", WebRequest.SCOPE_SESSION);
-        projects = projectService.getProjects(user);
-        tasks = taskService.getTasks(projectName, user);
-        model.addAttribute("tasks", tasks);
-        model.addAttribute("loopList", projects);
-        // model.addAttribute("project", project); Den her er ikke nødvendig. Den kender til den igennem ArrayListen med Projects :)
+        if (user == null){
+            return "login/login";
+        }
+        updateProjects(request, model);
         model.addAttribute("userName", user.getName());
         return "project/myProjects";// endpoint change
     }
@@ -44,16 +37,26 @@ public class ProjectController {
                 request.getParameter("start"),
                 request.getParameter("finish"),
                 Integer.parseInt(request.getParameter("budget")), user);
-        model.addAttribute("loopList", projects);
-        //model.addAttribute("userName", user.getName());
+        model.addAttribute("loopList", projects); //Skal den add projects her igen?
+        // Er det nødvendigt, når den adder i /myProjects, skal den ikke bare opdatere ArrayListen?
         return "redirect:/myProjects";
     }
 
     @GetMapping("/removeProject/{id}")
-    public String deleteProject(@PathVariable(value = "id") String id, Model model) {
+    public String deleteProject(@PathVariable(value = "id") String id, Model model, WebRequest request) {
+        if (user == null){
+            return "login/login";
+        }
         projectService.deleteProject(id, user);
-        projects = projectService.getProjects(user);
-        model.addAttribute("loopList", projects);
+        updateProjects(request,model);
         return "redirect:/myProjects";
     }
+
+    //Bruges til at opdatere Projects og adder til Model
+    public void updateProjects(WebRequest request, Model model){
+        user = (User) request.getAttribute("user", WebRequest.SCOPE_SESSION);
+        projects = projectService.getProjects(user);
+        model.addAttribute("loopList", projects);
+    }
+
 }
